backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy==2.0.35
alembic
pydantic
httpx
jinja2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
